chore: move node shapes into 'nodeShape' folder and rename them
