+++ client/src/components/nodeShapes/Ellipse.js
@@ -1,24 +1,24 @@
 import React from "react";
 import { Handle, Position } from "@xyflow/react";
 
-const roundedRectangleStyles = {
-  width: "150px",
-  height: "75px",
+const ellipseStyles = {
+  width: 150,
+  height: 75,
+  borderRadius: "50%",
   backgroundColor: "#eee",
   border: "2px solid #222",
-  borderRadius: "15px", // Rounded corners
   display: "flex",
   alignItems: "center",
   justifyContent: "center",
   textAlign: "center",
 };
 
-function RoundedRectangleNode() {
+function Ellipse() {
   return (
-    <div style={roundedRectangleStyles}>
+    <div style={ellipseStyles}>
       <Handle type="source" position={Position.Bottom} id="a" />
     </div>
   );
 }
 
-export default RoundedRectangleNode;
+export default Ellipse;
